refactor: member api response 타입 분리 및 memberApi prefix 네이밍을 http method명으로 통일
