+++ app/_apis/member.ts
@@ -1,30 +1,27 @@
-import { MemberResponse } from '@/_types/response';
+import type { MemberInfoParams, MemberResponse } from '@/_types/response/member';
 
 import instance from './core';
 
-export interface MemberInfoParams {
-  email: string;
-  nickname: string;
-}
+const MEMBER_API_URL = '/api/member';
 
 export const memberApi = {
   /**
    * @description 회원정보 조회
    */
-  getMember: () => instance.get<unknown, MemberResponse>(`/api/member`),
+  getMember: () => instance.get<unknown, MemberResponse>(MEMBER_API_URL),
   /**
    * @description 회원정보 등록
    */
-  registerMember: (info: MemberInfoParams) =>
-    instance.post<MemberInfoParams, MemberResponse>(`/api/member`, info),
+  postMember: (info: MemberInfoParams) =>
+    instance.post<MemberInfoParams, MemberResponse>(MEMBER_API_URL, info),
 
   /**
    * @description 회원정보 삭제
    */
-  deleteMember: () => instance.delete(`/api/member`),
+  deleteMember: () => instance.delete(MEMBER_API_URL),
   /**
    * @description 회원정보 수정
    */
-  updateMember: (info: MemberInfoParams) =>
-    instance.patch<MemberInfoParams, MemberResponse>(`/api/member`, info),
+  patchMember: (info: MemberInfoParams) =>
+    instance.patch<MemberInfoParams, MemberResponse>(MEMBER_API_URL, info),
 };
