+++ app/_types/response.ts
@@ -7,12 +7,6 @@ export interface AuthResponse {
   email: string;
 }
 
-export interface MemberResponse {
-  user_id: string;
-  email: string;
-  nickname: string;
-}
-
 export type APIErrorResponse = {
   status: HttpStatusCode;
   statusText: string;
