+++ app/_types/response/member.ts
@@ -1,0 +1,10 @@
+interface MemberInfo {
+  email: string;
+  nickname: string;
+}
+
+export type MemberInfoParams = MemberInfo;
+
+export interface MemberResponse extends MemberInfo {
+  user_id: string;
+}
